chore: adjust remove user feedback

+++ src/view/pages/Users/components/DeleteUserAction.tsx
@@ -1,7 +1,15 @@
-import React from 'react'
+import { useState } from 'react'
+import { Loader2 } from 'lucide-react'
 
 import { Button } from '@/view/components/ui/button'
-import { Alert } from '@/view/components/Alert'
+import {
+  AlertDialog,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+} from '@/view/components/ui/alert-dialog'
 
 interface DeleteUserActionProps {
   isOpen: boolean
@@ -10,23 +18,35 @@ interface DeleteUserActionProps {
 }
 
 export const DeleteUserAction = ({ isOpen, onDelete, onClose }: DeleteUserActionProps) => {
+  const [isLoading, setIsLoading] = useState(false)
+
   return (
-    <Alert
-      isOpen={isOpen}
-      title="Apagar usuário"
-      description="Deseja mesmo apagar o usuário?"
-      cancelButton={<Button variant="outline">Cancelar</Button>}
-      proceedButton={
-        <Button
-          variant="destructive"
-          onClick={async () => {
-            await onDelete()
-          }}
-        >
-          Apagar
-        </Button>
-      }
-      onClose={onClose}
-    />
+    <AlertDialog open={isOpen}>
+      <AlertDialogContent onKeyDown={(e) => e.stopPropagation()}>
+        <AlertDialogHeader>
+          <AlertDialogTitle>Apagar usuário</AlertDialogTitle>
+          <AlertDialogDescription>Deseja mesmo apagar o usuário?</AlertDialogDescription>
+        </AlertDialogHeader>
+        <AlertDialogFooter>
+          <Button variant="outline" disabled={isLoading} onClick={onClose} autoFocus>
+            Cancelar
+          </Button>
+
+          <Button
+            variant="destructive"
+            onClick={async () => {
+              setIsLoading(true)
+              await onDelete()
+              onClose()
+              setIsLoading(false)
+            }}
+            disabled={isLoading}
+          >
+            {isLoading && <Loader2 className="mr-2 w-4 h-4 animate-spin" />}
+            Apagar
+          </Button>
+        </AlertDialogFooter>
+      </AlertDialogContent>
+    </AlertDialog>
   )
 }
